+++ tests/test_torchnn.py
@@ -49,3 +49,25 @@ def test_quantumnet(backend):
     print(yp)
 
     np.testing.assert_allclose(yp.shape, np.array([3, n]))
+
+
+@pytest.mark.parametrize("backend", [lf("tfb"), lf("jaxb")])
+def test_inputs_multiple(backend):
+    n = 3
+    p = 0.1
+    K = tc.backend
+
+    def f(state, noise, weights):
+        c = tc.Circuit(n, inputs=state)
+        for i in range(n):
+            c.rz(i, theta=weights[i])
+        for i in range(n):
+            c.depolarizing(i, px=p, py=p, pz=p, status=noise[i])
+        return K.real(c.expectation_ps(x=[0]))
+
+    layer = tc.TorchLayer(f, [n], use_vmap=True, vectorized_argnums=[0, 1])
+    l = layer(
+        tc.get_backend("pytorch").ones([2, 2**n]) / 2 ** (n / 2),
+        0.2 * tc.get_backend("pytorch").ones([2, n], dtype="float32"),
+    )
+    print(l)